added selfie screen
I was able to get the map view to mostly work like I want. I want the
text to change when the map finishes animating. right now I just have a
3 second delay to simulate that. Hopefully tim can help me with that on
thursday.

The selfie view is mostly how I want. I wish that the view was just a
camera view but I couldn’t figure out how to do that. So instead, I
just preloaded 6 selfie jpg’s and presented buttons to take a selfie,
choose a selfie or skip the step. I wish I could animate the 6 images
but I didn’t want to waste time on it since I really just want it to
show a the view from the camera and let the “Take Picture” button be a
Shutter button.