A bunch of changes
I found a way to do the cool ios7 blur effect on our dialog at the
bottom. Its pretty sweet!

Also, I finally got the image to save to the disk and then load from
the disk. Also, I added an if statement so that if the device that
we’re running on doesn’t have a camera it will change the buttons so
there isn’t a take picture button.

You can comment out the if statement if you want to make the simulator
crash :p